+++ kugl/builtins/resources.py
@@ -1,4 +1,5 @@
 import json
+import re
 import sys
 from os.path import expandvars, expanduser
 from pathlib import Path
@@ -8,6 +9,7 @@ import yaml
 from pydantic import model_validator
 
 from kugl.api import resource, fail, run, Resource
+from kugl.util import best_guess_parse, KPath, debugging
 
 
 class NonCacheableResource(Resource):
@@ -41,12 +43,55 @@ class FileResource(NonCacheableResource):
 
     def get_objects(self):
         if self.file == "stdin":
-            return _parse(sys.stdin.read())
+            return best_guess_parse(sys.stdin.read())
         try:
             file = expandvars(expanduser(self.file))
-            return _parse(Path(file).read_text())
+            return KPath(file).parse()
         except OSError as e:
-            fail(f"failed to read {self.file}", e)
+            fail(f"failed to read {self.file} in resource {self.name}", e)
+
+
+@resource("folder")
+class FolderResource(NonCacheableResource):
+    """A resource that reads selectively from a folder tree.
+
+    These are non-cacheable for the same reason as FileResource."""
+    folder: Union[str, Path]
+    glob: str
+    match: str
+
+    @model_validator(mode="after")
+    @classmethod
+    def validate_folder(cls, resource: "FolderResource"):
+        try:
+            resource._pattern = re.compile(resource.match)
+        except Exception as e:  # re.compile can raise anything
+            fail(f"Invalid regex {resource.match} in resource {resource.name}")
+        if isinstance(resource.folder, str):
+            resource.folder = KPath(expandvars(expanduser(resource.folder)))
+        if not resource.folder.exists():
+            fail(f"Missing resource folder {resource.folder}")
+        return resource
+
+    def get_objects(self):
+        folder = KPath(expandvars(expanduser(str(self.folder))))
+        files = [p.relative_to(folder) for p in folder.glob(self.glob)]
+        if not files:
+            fail(f"Glob {self.glob} in {folder} produced no files")
+        result = []
+        debug = debugging("folder")
+        if debug:
+            debug(f"Reviewing files for {self.glob} in {folder}")
+        for file in files:
+            m = self._pattern.search(str(file))
+            if m:
+                if debug:
+                    debug(f"Adding {file} with match {m.groupdict()}")
+                result.append(dict(content=folder.joinpath(file).parse(), match=m.groupdict()))
+            else:
+                if debug:
+                    debug(f"Skipping {file}, did not match regex")
+        return result
 
 
 @resource("exec")
@@ -70,17 +115,8 @@ class ExecResource(Resource):
 
     def get_objects(self):
         _, out, _ = run(self.exec)
-        return _parse(out)
+        return best_guess_parse(out)
 
     def cache_path(self):
         assert self.cache_key is not None  # should be covered by validator
         return f"{expandvars(self.cache_key)}/{self.name}.exec.json"
-
-
-def _parse(text):
-    if not text:
-        return {}
-    if text[0] in "{[":
-        return json.loads(text)
-    return yaml.safe_load(text)
-