+++ kugl/util/misc.py
@@ -9,6 +9,7 @@ from contextlib import contextmanager
 from typing import Optional, Union, Callable, Tuple
 
 import arrow
+import yaml
 
 from .debug import debugging
 
@@ -83,3 +84,23 @@ def cleave(s: str, sep: str, flip: bool = False):
         parts = s.split(sep, 1)
         return parts[0], parts[1]
     return (None, s) if flip else (s, None)
+
+
+def friendlier_errors(errors: list) -> list[str]:
+    """Improve upon Pydantic's error messages."""
+    location_str = lambda loc: ".".join(map(str, loc))
+    def _improve(error):
+        message, location = error['msg'], error['loc']
+        if "Extra inputs are not permitted" in message:
+            return f"At {location_str(location[:-1])}: '{location[-1]}' is not allowed here"
+        return location_str(location) + ": " + message
+    return [_improve(e) for e in errors]
+
+
+def best_guess_parse(text):
+    if not text:
+        return {}
+    if text[0] in "{[":
+        return json.loads(text)
+    return yaml.safe_load(text)
+