+++ tests/config/test_merge_schema.py
@@ -1,41 +1,17 @@
 """
-Unit tests for multiple config files on init_path.
+Unit tests for multiple schema config files on init_path.
 """
-
-import sqlite3
 
 import pytest
 
 from kugl.main import main1
-from kugl.util import KPath, kugl_home, KuglError
-from tests.testing import assert_by_line
+from kugl.util import kugl_home, KuglError
+from ..testing import assert_by_line
 
 
-@pytest.fixture(scope="function")
-def extra_home(test_home, tmp_path):
-    """Additional home for Kugl init and schema files.
-
-    When used, this fixture creates a second folder similar to test_home, and writes init.yaml
-    in test_home pointing to both folders."""
-    kugl_home().prep().joinpath("init.yaml").write_text(f"""
-        settings:
-          init_path:
-            - "{kugl_home()}"
-            - "{tmp_path}"
-    """)
-    yield KPath(tmp_path).prep()
-
-
-def test_bogus_init_paths(hr, test_home):
+def test_bogus_init_paths(hr, extra_home):
     """Put only a missing folder in the init path, verify that breaks."""
-    hr.save()
-    # Don't use the extra_home fixture, because don't want to see both entries in the init path
-    bad_home = test_home / "xyz"
-    kugl_home().prep().joinpath("init.yaml").write_text(f"""
-        settings:
-          init_path:
-            - {bad_home}
-    """)
+    extra_home.rmdir()
     with pytest.raises(KuglError, match="no configurations found for schema 'hr'"):
         main1([hr.PEOPLE_QUERY])
 
@@ -46,7 +22,7 @@ def test_reject_dupe_resource(hr, extra_home):
     extra_home.joinpath("hr.yaml").write_text("""
         resources:
         - name: people
-          data: []
+          data: {}
     """)
     with pytest.raises(KuglError, match="Resource 'people' is already defined in schema 'hr'"):
         main1([hr.PEOPLE_QUERY])
@@ -54,8 +30,8 @@ def test_reject_dupe_resource(hr, extra_home):
 
 def test_reject_dupe_table(hr, extra_home):
     """Table must not be defined in more than one schema file"""
-    hr.save()
-    extra_home.joinpath("hr.yaml").write_text("""
+    hr.save(folder=extra_home)
+    kugl_home().joinpath("hr.yaml").write_text("""
         create:
         - table: people
           resource: people
@@ -69,8 +45,8 @@ def test_reject_dupe_table(hr, extra_home):
 
 def test_reject_dupe_column(hr, extra_home):
     """Column must not be defined in more than one schema file"""
-    hr.save()
-    extra_home.joinpath("hr.yaml").write_text("""
+    hr.save(folder=extra_home)
+    kugl_home().joinpath("hr.yaml").write_text("""
         extend:
         - table: people
           columns:
@@ -83,8 +59,8 @@ def test_reject_dupe_column(hr, extra_home):
 
 def test_extend_valid_table(hr, extra_home, capsys):
     """Verify result of extending a table from a separate schema file."""
-    hr.save()
-    extra_home.joinpath("hr.yaml").write_text("""
+    hr.save(folder=extra_home)
+    kugl_home().joinpath("hr.yaml").write_text("""
         extend:
         - table: people
           columns: