+++ tests/test_cli.py
@@ -59,22 +59,13 @@ def test_unknown_option(test_home, capsys):
 def test_enforce_one_cache_option_via_shortcut(test_home, capsys):
     kugl_home().prep().joinpath("init.yaml").write_text("""
         shortcuts:
-          foo:
-          - -u
-          - "select 1"
+          - name: foo
+            args:
+              - -u
+              - "select 1"
     """)
     with pytest.raises(KuglError, match="Cannot use both -c/--cache and -u/--update"):
         main1(["-c", "foo"])
-
-
-def test_simple_shortcut(test_home, capsys):
-    kugl_home().prep().joinpath("init.yaml").write_text("""
-        shortcuts:
-          foo: ["select 1, 2"]
-    """)
-    main1(["foo"])
-    out, _ = capsys.readouterr()
-    assert out == "  1    2\n" * 2
 
 
 def test_no_headers(test_home, capsys):